Re-imports the tests from svn, added makefile targets
These tests have not been run since 2010.

So we had to do some archeology to see which version these
seem to be made for. I think back then we ran the tests
differently, e.g. something like php AllTests.php, but it
seems to work like this.

Also, back then there was no composer or such, so we do the
importing of dependencies the old way. Luckily there's only
two depencies.